L3.Core

Description

Type checking and type inference

Documentation

type ShowCtx = Context Name

type ShowExpr = Expr Name

newtype Name

Constructors

Name String

type Context a = [(a, Expr a)]

A context is a stack, mapping names to bound values.

data Expr a

An expression in the calculus of constructions.

Constructors

Star

Box

Var a

Lam a (Expr a) (Expr a)

Pi a (Expr a) (Expr a)

App (Expr a) (Expr a)

showExpr :: ShowExpr -> String

Show an expression

showCtx :: Show a => Context a -> String

Show for a context, printing each binding on a separate line.

debugCore :: String -> a -> a

free :: (Eq a, Show a) => a -> Expr a -> Bool

Is a name free in an expression In this context, free v a & v = v' => substitute v v' a = a i.e. would a substitution be performed

free' :: (Eq a, Show a) => a -> Expr a -> Bool

fresh :: (Eq a, Enum a, Show a) => a -> Expr a -> a

fresh' :: (Eq a, Show a, Enum a) => a -> Expr a -> a

substitute :: (Eq a, Enum a, Show a) => a -> Expr a -> Expr a -> Expr a

Substitute all occurrences of a variable v with an expression e. substitute v e E ~ E[v := e]

substitute' :: (Eq a, Enum a, Show a) => a -> Expr a -> Expr a -> Expr a

normalize :: (Eq a, Enum a, Show a) => Expr a -> Expr a

Given an expression, reduce it one step towards its normal form.

normalize' :: (Eq a, Enum a, Show a) => Expr a -> Expr a

normalize0 :: (Eq a, Enum a, Show a) => Expr a -> Expr a

Given an expression, totally reduce it over all steps towards normal form, returning this normal.

normalize0' :: (Eq a, Enum a, Show a) => Expr a -> Expr a

index :: (Eq a, Enum a, Show a) => Int -> Expr (Either Int a) -> Expr (Either Int a)

Given an free index, convert an expression in Right names into Left indexes. This uses DeBruijn indicies.

index' :: (Eq a, Enum a, Show a) => Int -> Expr (Either Int a) -> Expr (Either Int a)

index0 :: (Eq a, Enum a, Show a) => Expr a -> Expr (Either Int a)

Provide an initial free index of 0 and index an expression. This converts any expression to its DeBruijn indexed form, leaving global names untouched.

index0' :: (Eq a, Enum a, Show a) => Expr a -> Expr (Either Int a)

alphaEq :: (Eq a, Enum a, Show a) => Expr a -> Expr a -> Bool

Deduce whether two expressions are equivalent by converting to indexed form and checking for exact equality. This does not apply normalisation, so represents only alpha-equivalence of expressions.

alphaEq' :: (Eq a, Enum a, Show a) => Expr a -> Expr a -> Bool

betaEq :: (Eq a, Enum a, Show a) => Expr a -> Expr a -> Bool

Deduce whether two expressions are equivalent by converting to indexed form and checking for exact equality. This does apply normalisation, so represents beta-equivalence (and implicitly alpha-equivalence) of expressions.

betaEq' :: (Eq a, Enum a, Show a) => Expr a -> Expr a -> Bool

evalExpr :: (Eq a, Enum a, Show a) => Context a -> Expr a -> Result (Expr a, Expr a)

Evaluate an expression, returning its type and normalized form

evalExpr' :: (Eq a, Enum a, Show a) => Context a -> Expr a -> Result (Expr a, Expr a)

inferType1 :: (Eq a, Enum a, Show a) => Context a -> Expr a -> Result (Expr a)

Type-check an expression and return the expression's type if type-checking succeeds or an error message if type-checking fails inferType` does not necessarily normalize the type since full normalization is not necessary for just type-checking. If you actually care about the returned type then you may want to normalize it afterwards. Type inference is within a type context (list of global names and their types)

inferType1' :: (Show a, Eq a, Enum a) => [(a, Expr a)] -> Expr a -> Either Error (Expr a)

inferType :: (Eq a, Enum a, Show a) => Context a -> Expr a -> Result (Expr a)

Type-check an expression and return the expression's normalized type if type-checking succeeds or an error message if type-checking fails

inferType' :: (Eq a, Enum a, Show a) => Context a -> Expr a -> Result (Expr a)

inferType0 :: (Eq a, Enum a, Show a) => Expr a -> Result (Expr a)

inferType0 is the same as inferType with an empty context, meaning that the expression must be closed (i.e. no free variables), otherwise type-checking will fail.

inferType0' :: (Eq a, Enum a, Show a) => Expr a -> Result (Expr a)

wellTyped :: (Eq a, Enum a, Show a) => Context a -> Expr a -> Bool

Deduce if an expression e is well-typed - i.e. its type can be inferred.

wellTyped' :: (Eq a, Enum a, Show a) => Context a -> Expr a -> Bool

wellTyped0 :: (Eq a, Enum a, Show a) => Expr a -> Bool

Deduce if an expression is well-typed context-free - i.e. it is additionally closed and therefore well-typed without additional context.

wellTyped0' :: (Eq a, Enum a, Show a) => Expr a -> Bool

module L3.Util

Orphan instances

(Enum a, Enum b) => Enum (Either a b)

Instance details

Methods

succ :: Either a b -> Either a b

pred :: Either a b -> Either a b

toEnum :: Int -> Either a b

fromEnum :: Either a b -> Int

enumFrom :: Either a b -> [Either a b]

enumFromThen :: Either a b -> Either a b -> [Either a b]

enumFromTo :: Either a b -> Either a b -> [Either a b]

enumFromThenTo :: Either a b -> Either a b -> Either a b -> [Either a b]