L3.Parse.Lexer

Description

Lexer from Strings into Tokens

Documentation

data Token

Lex strings into tokens, canonical form for the syntax definitition

Constructors

OpenParen

CloseParen

OpenBracket

CloseBracket

HasType

At

StarT

BoxT

Arrow

LambdaT

PiT

AutoT

Symbol String

Number Int

Comment String

EOL

alternatives :: [Parser String Token]

A list of Alternatives that may be used to lex a string into tokens.

comment :: Parser String Token

Parse a comment, taking characters until an End-Of-Line

lexSrc :: String -> Result [Token]

Parse a string into canonical form using tokens

grammar :: Parser String [Token]

The grammar for this parser is the collection of Alternatives