L3.Parser

Description

Parser from Tokens into Expressions

Documentation

debugParser :: String -> a -> a

debugParserM :: (i -> String) -> Parser i o -> Parser i o

parseExpr :: [Token] -> Result ShowExpr

Parse a string to a named expression (using string labels) Strip comments here, as they apply inside any context and are difficult to deal with otherwise

parseExpr' :: [Token] -> Result ShowExpr

sugarE :: Parser [Token] ShowExpr

Sugared expression, injection point for additional syntax niceties S :: [π(_:]A[)].S | A

sugarE' :: Parser [Token] ShowExpr

appE :: Parser [Token] ShowExpr

Applicative expression, unknown and unbounded length A :: F [app F ..]

appE' :: Parser [Token] (Expr Name)

funE :: Parser [Token] ShowExpr

Function expression F :: λ(s:τ).E | π(s:τ).E | T | (S)

funE' :: Parser [Token] ShowExpr

termE :: Parser [Token] ShowExpr

Terminal expression, bounded in length and with no children T :: * | # | n@v | v

termE' :: Parser [Token] ShowExpr

arrE :: Parser [Token] (Name, ShowExpr)

Arrow expression, a component of functions (->) :: (s:τ).

arrE' :: Parser [Token] (Name, ShowExpr)

typE :: Parser [Token] (Name, ShowExpr)

Type expression, a symbol has type expr τ :: s:S

typE' :: Parser [Token] (Name, ShowExpr)

star :: Parser [Token] ShowExpr

Star axiom * :: Star

star' :: Parser [Token] (Expr a)

box :: Parser [Token] ShowExpr

Box axiom # :: Box

box' :: Parser [Token] (Expr a)

var :: Parser [Token] ShowExpr

Variable axiom s :: Var s

var' :: Parser [Token] (Expr Name)

nsVar :: Parser [Token] ShowExpr

Namespaced-variable axiom ns :: Var ns

nsVar' :: Parser [Token] (Expr Name)

lamE :: Parser [Token] ShowExpr

Lambda function λ(s:S).S :: Lam s S S

lamE' :: Parser [Token] (Expr Name)

piE :: Parser [Token] ShowExpr

Pi function π(s:S).S :: Lam s S S

piE' :: Parser [Token] (Expr Name)

anonPiE :: ShowExpr -> Parser [Token] ShowExpr

Anonymous-pi function, where the name is unused and therefore not written Note the implicit expression argument, as otherwise sugar = (app >> arrow >> sugar) | app If this were not the case, all expressions are parsed as anonymous pis and will fail at the final arrow This is an exponential slowdown S. :: π(_:S).S

anonPiE' :: Expr Name -> Parser [Token] (Expr Name)

module L3.Lexer

module L3.Core